# Conflicts:
#	.travis.yml
#	README.md
#	composer.json
#	extension/ezformtoken/event/ezxformtoken.php
#	extension/ezoe/design/standard/javascript/classes/dom/Selection.js
#	extension/ezoe/design/standard/javascript/tiny_mce.js
#	extension/ezoe/design/standard/javascript/tiny_mce_jquery.js
#	extension/ezoe/design/standard/javascript/tiny_mce_jquery_src.js
#	extension/ezoe/design/standard/javascript/tiny_mce_prototype.js
#	extension/ezoe/design/standard/javascript/tiny_mce_prototype_src.js
#	extension/ezoe/design/standard/javascript/tiny_mce_src.js
#	extension/ezoe/design/standard/templates/ezoe/tag_custom.tpl
#	extension/ezoe/ezxmltext/handlers/input/ezoexmlinput.php
#	extension/ezoe/modules/ezoe/classes/GoogleSpell.php
#	extension/ezoe/modules/ezoe/classes/SpellChecker.php
#	extension/ezoe/modules/ezoe/classes/utils/mcejson.php
#	extension/ezoe/modules/ezoe/spellcheck_rpc.php

+++ kernel/private/classes/clusterfilehandlers/dfsbackends/mysqli.php
@@ -334,7 +334,6 @@ class eZDFSFileHandlerMySQLiBackend implements eZClusterEventNotifier
      * @see _purge
      *
      * @return bool|int false if it fails, number of affected rows otherwise
-     * @todo This method should also remove the files from disk
      */
     public function _purgeByLike( $like, $onlyExpired = false, $limit = 50, $expiry = false, $fname = false )
     {
@@ -392,11 +391,16 @@ class eZDFSFileHandlerMySQLiBackend implements eZClusterEventNotifier
             return $this->_fail( "Purging file metadata by like statement $like failed" );
         }
         $deletedDBFiles = mysqli_affected_rows( $this->db );
-        $this->dfsbackend->delete( $files );
 
-        $this->_commit( $fname );
+        $commitResult = $this->_commit( $fname );
+        if ( $commitResult === true || $commitResult === null )
+        {
+            $this->dfsbackend->delete( $files );
 
-        return $deletedDBFiles;
+            return $deletedDBFiles;
+        }
+
+        return false;
     }
 
     /**
@@ -651,7 +655,7 @@ class eZDFSFileHandlerMySQLiBackend implements eZClusterEventNotifier
      * @param string $dir
      * @return bool
      */
-    protected function __mkdir_p( $dir )
+    private function mkdir_p( $dir )
     {
         // create parent directories
         $dirElements = explode( '/', $dir );
@@ -679,6 +683,16 @@ class eZDFSFileHandlerMySQLiBackend implements eZClusterEventNotifier
         }
 
         return $result;
+    }
+
+    /**
+     * @deprecated Use eZDFSFileHandlerMySQLiBackend::mkdir_p() instead
+     * @param $dir
+     * @return bool
+     */
+    protected function __mkdir_p( $dir )
+    {
+        return $this->mkdir_p( $dir );
     }
 
     /**
@@ -712,7 +726,7 @@ class eZDFSFileHandlerMySQLiBackend implements eZClusterEventNotifier
                 $tmpFilePath = substr_replace( $filePath, $tmpid, strrpos( $filePath, '.' ), 0  );
             else
                 $tmpFilePath = $filePath . '.' . $tmpid;
-            $this->__mkdir_p( dirname( $tmpFilePath ) );
+            $this->mkdir_p( dirname( $tmpFilePath ) );
             eZDebugSetting::writeDebug( 'kernel-clustering', "copying DFS://$filePath to FS://$tmpFilePath on try: $loopCount " );
 
             // copy DFS file to temporary FS path
@@ -1351,6 +1365,10 @@ class eZDFSFileHandlerMySQLiBackend implements eZClusterEventNotifier
     /**
      * Stops a current transaction and commits the changes by executing a COMMIT call.
      * If the current transaction is a sub-transaction nothing is executed.
+     *
+     * Returns true if the commit was executed successfully, false if it failed, null if it wasn't executed.
+     *
+     * @return bool|null
      */
     protected function _commit( $fname = false )
     {
@@ -1360,7 +1378,11 @@ class eZDFSFileHandlerMySQLiBackend implements eZClusterEventNotifier
             $fname = "_commit";
         $this->transactionCount--;
         if ( $this->transactionCount == 0 )
-            $this->_query( "COMMIT", $fname );
+        {
+            return $this->_query("COMMIT", $fname) !== false;
+        }
+
+        return null;
     }
 
     /**
